CRUD Course done, CRU Chapter done, CR Comment done

+++ database/migrations/2023_01_13_111508_create_courses_chapters_table.php
@@ -14,25 +14,25 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('courses_chapters', function (Blueprint $table) {
-            $table->id();
+        if (!Schema::hasTable('courses_chapters')) {
+            Schema::create('courses_chapters', function (Blueprint $table) {
+                $table->id();
 
-            $table->bigInteger('id_course')->unsigned();
-            $table->foreign('id_course')
-                ->references('id')
-                ->on('courses')
-                ->constrained()
-                ->onUpdate('cascade')
-                ->onDelete('cascade');
+                $table->bigInteger('id_course')->unsigned();
+                $table->foreign('id_course')
+                    ->references('id')
+                    ->on('courses')
+                    ->constrained()
+                    ->onDelete('cascade');
 
-            $table->bigInteger('id_chapter')->unsigned();
-            $table->foreign('id_chapter')
-                ->references('id')
-                ->on('chapters')
-                ->constrained()
-                ->onUpdate('cascade')
-                ->onDelete('cascade');
-        });
+                $table->bigInteger('id_chapter')->unsigned();
+                $table->foreign('id_chapter')
+                    ->references('id')
+                    ->on('chapters')
+                    ->constrained()
+                    ->onDelete('cascade');
+            });
+        }
 
         /*DB::table('courses_chapters')->insert([
             'id_course' => 1,
